refactored script and script engine member variables
In keeping with modern standards, class member variables should always be private with getter/setter where necessary and the variables should have either a prefix or suffix unique to private variables so that it does not conflict with publicly exposed functions/variables. I chose to suffix the variables with `_` i.e. `runtime_`.