refactored entry startup/shutdown into D2BS class
Big part of the refactoring of D2BS is untangling all the different free-functions into classes so that the global namespace is not cluttered with functions that is called from absolutely everywhere.

The main entry point goes directly into the new `D2BS` class which is the heart and soul of the program.