new class Application replaces Engine
The new Application class will act sort of like the runtime of the framework, it initializes, runs then shuts down while dealing with errors occuring during any of those stages.

DllMain now only spawns a thread that does the most basic setup, taking the input HMODULE and grabbing the path of our DLL which acts as our working directory. After that it creates a new instance of Application, runs then reset once run returns to call the destructor which is replaces shutdown calls.

Engine.h is still present for the includes until it gets cleaned up.